fw-5518, extract a base class for entry search (cross-site and single-site) (#1255)
* fw-5518, extract search entries base class

- put common params in the base class
- put custom param support in the correct APIs
- no functional changes

* fw-5518, tests for search result serialization (#1256)

- convert test_hydrate_objects into a mixin to test serialization of search results in the views
