+++ firstvoices/backend/tests/test_apis/test_search_apis/test_site_search_api.py
@@ -7,14 +7,23 @@ from elasticsearch_dsl import Search
 from backend.models.constants import Role, Visibility
 from backend.tests import factories
 from backend.tests.test_apis.base_api_test import BaseSiteContentApiTest
+from backend.tests.test_apis.test_search_apis.base_search_test import SearchMocksMixin
+from backend.tests.test_apis.test_search_apis.test_search_querying.test_search_entry_results import (
+    SearchEntryResultsTestMixin,
+)
 
 
 @pytest.mark.django_db
-class TestSiteSearchAPI(BaseSiteContentApiTest):
+class TestSiteSearchAPI(
+    SearchEntryResultsTestMixin, SearchMocksMixin, BaseSiteContentApiTest
+):
     """Remaining tests that cover the site search."""
 
     API_LIST_VIEW = "api:site-search-list"
     API_DETAIL_VIEW = "api:site-search-detail"
+
+    def get_search_endpoint(self, site):
+        return f"{self.get_list_endpoint(site_slug=site.slug)}?q=what"
 
     def test_invalid_category_id(self):
         site, user = factories.get_site_with_member(
@@ -52,7 +61,7 @@ class TestSiteSearchAPI(BaseSiteContentApiTest):
         self.client.force_authenticate(user=user)
 
         with patch(
-            "backend.views.search_all_entries_views.get_search_query",
+            "backend.views.base_search_entries_views.get_search_query",
             return_value=Search(),
         ) as mock_get_search_query:
             self.client.get(
